refactor: change absolute path to relative
