+++ apps/shared-models/speaker_resource.model.ts
@@ -1,18 +1,3 @@
-import { IUser } from './user.model';
-import { IEvent } from './event.model';
-import { ITag } from 'apps/shared-models/tag.model';
+import { ISpeakerResource } from '@commudle/shared-models';
 
-export interface ISpeakerResource {
-  id: number;
-  title: string;
-  session_details_links: string;
-  embedded_content: string;
-  user: IUser;
-  event: IEvent;
-  created_at: Date;
-  user_event_registration_id: number;
-  data_form_entity_response_group_id: number;
-  tags?: ITag[];
-  favicon?: string;
-  description?: string;
-}
+export { ISpeakerResource };
