+++ libs/shared/components/src/index.ts
@@ -3,3 +3,7 @@ export * from './lib/shared-components.module';
 export * from './lib/components/discussion/discussion.component';
 
 export * from './lib/components/vote/vote.component';
+
+export * from './lib/components/channel-discussion/channel-discussion.component';
+
+export * from './lib/components/loading-spinner/loading-spinner.component';
