--- libs/shared/models/src/lib/user_roles.enum.ts
@@ -1,15 +1,0 @@
-export enum EUserRoles {
-  ORGANIZER = 'organizer',
-  EVENT_ORGANIZER = 'event_manager',
-  COMMUNITY_ADMIN = 'community_administrator',
-  MEMBER = 'member',
-  EVENT_VOLUNTEER = 'event_volunteer',
-  SYSTEM_ADMINISTRATOR = 'system_administrator',
-  COMMUNITY_CHANNEL_ADMIN = 'community_channel_admin',
-  COMMUNITY_CHANNEL_MEMBER = 'community_channel_member',
-  NEWSLETTER = 'newsletter',
-  PAGE_ADS = 'page_ads',
-  BADGES = 'badges',
-  FEATURED_COMMUNITIES = 'featured_communities',
-  STATIC_ASSETS = 'static_assets',
-}
